+++ src/commands.js
@@ -1,3 +1,4 @@
+import "./commands/fill-form-field";
 import "./commands/get-by-data";
 import "./commands/get-form-field";
 import "./commands/should-be-visible";
@@ -5,6 +6,7 @@ import "./commands/should-have-attribute";
 import "./commands/should-have-class";
 import "./commands/should-have-count";
 import "./commands/should-have-text";
+import "./commands/should-have-value";
 import "./commands/should-not-be-visible";
 import "./commands/should-not-have-attribute";
 import "./commands/should-not-have-class";